# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Albelli API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/advertisements": {
            "get": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Get all the available advertisements",
                "operationId": "4053df07b4f05b7c22217d739f1956f3",
                "responses": {
                    "200": {
                        "description": "A list of advertisements.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseAdvertisement"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Adds a new advertisement",
                "operationId": "fa6fb514310a3528c4279ecf7a715e26",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdvertisementCreateSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AdvertisementResource created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseAdvertisement"
                                },
                                "example": {
                                    "success": true,
                                    "message": "AdvertisementResource My title created.",
                                    "data": {
                                        "id": 12,
                                        "title": "My title",
                                        "created_at": "2021-09-26T18:17:47+00:00",
                                        "updated_at": "2021-09-26T18:17:47+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": {
                                        "title": [
                                            "The title field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/advertisements/{id}": {
            "get": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Returns a specific instance given by the id",
                "operationId": "ad4cfb43c9744b2cc406a257a62dddb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retuns one instance with the asked advertisement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseAdvertisement"
                                },
                                "example": {
                                    "success": true,
                                    "message": "AdvertisementResource My title fetched.",
                                    "data": {
                                        "id": 12,
                                        "title": "My title",
                                        "created_at": "2021-09-26T18:17:47+00:00",
                                        "updated_at": "2021-09-26T18:17:47+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AdvertisementResource does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "AdvertisementResource does not exist."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Updates the existing advertisement given by the id",
                "operationId": "5025f58b7c89dff2898583502e3405a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdvertisementCreateSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AdvertisementResource updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseAdvertisement"
                                },
                                "example": {
                                    "success": true,
                                    "message": "AdvertisementResource My title3 updated.",
                                    "data": {
                                        "id": 1,
                                        "title": "My title3",
                                        "created_at": "2021-09-26T16:08:55+00:00",
                                        "updated_at": "2021-09-26T18:40:53+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": {
                                        "title": [
                                            "The title field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Advertisements"
                ],
                "summary": "Deletes the advertisement given by the id",
                "operationId": "baa40576659cae6050c3b7c6bf0612f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdvertisementResource deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "AdvertisementResource MyTitle deleted."
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Error Referenced",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Cannot proceed with query, it is referenced by other records in the database."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error Not exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "AdvertisementResource does not exist."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/offers": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "summary": "Get all the available offers",
                "operationId": "35f8b3321cde4a9a62e4f052979ef3e8",
                "responses": {
                    "200": {
                        "description": "A list of offers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseOffer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Offers"
                ],
                "summary": "Adds a new offer",
                "operationId": "dd33d47e1d3fe777b736a1b6d711779d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfferCreateSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OfferResource created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseOffer"
                                },
                                "example": {
                                    "success": true,
                                    "message": "OfferResource The product name created.",
                                    "data": {
                                        "id": 14,
                                        "product_name": "The product name",
                                        "discount_value": 25,
                                        "start_date": "2021-09-26T10:57:37+00:00",
                                        "end_date": "2021-09-26T10:57:37+00:00",
                                        "created_at": "2021-09-26T20:47:39+00:00",
                                        "updated_at": "2021-09-26T20:47:39+00:00"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": {
                                        "discount_value": [
                                            "The discount value field is required."
                                        ],
                                        "start_date": [
                                            "The start date field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "JsonResponse": {
                "title": "Json Response Model",
                "description": "Json Response Model",
                "properties": {
                    "success": {
                        "description": "Returns true if the action was completed succesfully",
                        "type": "bolean",
                        "example": true
                    },
                    "message": {
                        "description": "Includes any aditional message that the endpoint may generate with informative proposes",
                        "type": "string",
                        "example": "Inserted ok. The query took 100ms"
                    },
                    "data": {
                        "description": "Optional data returned by the API server goes here or an error message if success was false"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponse"
                }
            },
...
            "AdvertisementCreateSchema": {
                "title": "AdvertisementResource Create Update Model",
                "description": "AdvertisementResource Create Update Model",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the advertisement",
                        "type": "string",
                        "example": "My title"
                    },
                    "offers": {
                        "description": "[OPTIONAL] The ids of the related offers. All ids should exists before creating this advertisement",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": "[1,2,3]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AdvertisementCreateSchema"
                }
            },
            "JsonResponseAdvertisement": {
                "title": "Json Response AdvertisementResource Model",
                "description": "Json AdvertisementResource Model",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/JsonResponse"
                    },
                    {
                        "required": [
                            "success",
                            "message"
                        ],
                        "properties": {
                            "success": {
                                "description": "Returns true if the action was completed succesfully",
                                "type": "bolean",
                                "example": true
                            },
                            "message": {
                                "description": "Includes any aditional message that the endpoint may generate with informative proposes",
                                "type": "string",
                                "example": "Inserted ok. The query took 100ms"
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AdvertisementSchema"
                                }
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "JsonResponseAdvertisement"
                        }
                    }
                ]
            },
...
            "JsonResponseOffer": {
                "title": "Json Response OfferResource Model",
                "description": "Json OfferResource Model",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/JsonResponse"
                    },
                    {
                        "required": [
                            "success",
                            "message"
                        ],
                        "properties": {
                            "success": {
                                "description": "Returns true if the action was completed succesfully",
                                "type": "bolean",
                                "example": true
                            },
                            "message": {
                                "description": "Includes any aditional message that the endpoint may generate with informative proposes",
                                "type": "string",
                                "example": "Inserted ok. The query took 100ms"
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OfferSchema"
                                }
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "JsonResponseOffer"
                        }
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}